requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
django==5.0.3; python_version >= '3.10'
djangorestframework==3.14.0; python_version >= '3.6'
drf-spectacular==0.27.1; python_version >= '3.6'
drf-spectacular-sidecar==2024.3.4; python_version >= '3.6'
gunicorn==21.2.0; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.33.0; python_version >= '3.8'
rpds-py==0.18.0; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2024.1; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
